\name{Copula.Markov-package}
\alias{Copula.Markov-package}
\alias{Copula.Markov}
\docType{package}
\title{ Copula-Based Estimation and Statistical Process Control for Serially Correlated Time Series }
\description{
Copulas are applied to model a Markov dependence for serially correlated time series.
 The Clayton and Joe copulas are available to specify the dependence structure.
 The normal and binomial distributions are available for the marginal model.
 Maximum likelihood estimation is implmented for estimating parameters,
 and a Shewhart control chart is drawn for performing statistical process control.
}
\details{
\tabular{ll}{
Package: \tab Copula.Markov\cr
Type: \tab Package\cr
Version: \tab 2.6\cr
Date: \tab 2020-1-8\cr
License: GPL-2
}
}
\author{
Emura T, Chen WR, Long TH, Huang X.
 Maintainer: Takeshi Emura <takeshiemura@gmail.com>
}
\references{
Chen W (2018) Copula-based Markov chain model with binomial data, NCU Library

Huang XW, Chen W, Emura T (2019-), Likelihood-based inference for a copula-based Markov
 chain model with binomial time series, in review

Emura T, Long TH, Sun LH (2017), R routines for performing estimation and
 statistical process control under copula-based time series models,
 Communications in Statistics - Simulation and Computation, 46 (4): 3067-87

Long TH and Emura T (2014), A control chart using copula-based Markov chain models,
 Journal of the Chinese Statistical Association 52 (No.4): 466-96

Huang XW, Emura T (2019), Model diagnostic procedures for copula-based Markov chain models
 for statistical process control, Communications in Statistics - Simulation and Computation,
 doi:10.1080/03610918.2019.1602647

}
\keyword{ package }
