\name{Copula.Markov-package}
\alias{Copula.Markov-package}
\alias{Copula.Markov}
\docType{package}
\title{
Estimation and Statistical Process Control Under Copula-Based Time Series Models
}
\description{
Estimation and statistical process control are performed under copula-based time-series models.
}
\details{
\tabular{ll}{
Package: \tab Copula.Markov\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2016-07-30\cr
License: GPL-2
}
}
\author{
Takeshi Emura <takeshiemura@gmail.com>
}
\references{
Emura T, Long TH, Sun LH (2015), R routines for performing estimation and statistical process control under copula-based time series models, 
Communications in Statistics - Simulation and Computation, DOI: 10.1080/03610918.2015.1073303

Long TH and Emura T (2014), A control chart using copula-based Markov chain models, Journal of the Chinese Statistical Association 52 (No.4): 466-96
}
\keyword{ package }
