\name{Copula.Markov-package}
\alias{Copula.Markov-package}
\alias{Copula.Markov}
\docType{package}
\title{ Copula-Based Estimation and Statistical Process Control for Serially Correlated Time Series }
\description{
Copulas are applied to model serial dependence in time series.
 The Clayton and Joe copulas are available for the dependence structure.
 The normal and binomial distributions are available for the marginal model.
}
\details{
\tabular{ll}{
Package: \tab Copula.Markov\cr
Type: \tab Package\cr
Version: \tab 2.1\cr
Date: \tab 2018-07-23\cr
License: GPL-2
}
}
\author{
Takeshi Emura <takeshiemura@gmail.com>
}
\references{
Chen W (2018) Copula-based Markov chain model with binomial data, NCU Library

Chen W, Emura T (2018-), A control chart using a copula-based Markov chain
 for attribute data, submitted for publication

Emura T, Long TH, Sun LH (2017), R routines for performing estimation and
 statistical process control under copula-based time series models,
 Communications in Statistics - Simulation and Computation, 46 (4): 3067-87

Long TH and Emura T (2014), A control chart using copula-based Markov chain models,
 Journal of the Chinese Statistical Association 52 (No.4): 466-96
}
\keyword{ package }
