\name{Copula.Markov-package}
\alias{Copula.Markov-package}
\alias{Copula.Markov}
\docType{package}
\title{
Estimation and Statistical Process Control Under Copula-Based Time Series Models
}
\description{
Estimation and statistical process control are performed under copula-based time-series models.
}
\details{
\tabular{ll}{
Package: \tab Copula.Markov\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-23\cr
License: GPL-2
}
}
\author{
Takeshi Emura <emura@stat.ncu.edu.tw>
}
\references{
Emura T and Long TH (2015), R routines for performing estimation and statistical process control under copula-based time series models, 
under revision, Communications in Statistics - Simulation and Computation

Long TH and Emura T (2014), A control chart using copula-based Markov chain models, Journal of the Chinese Statistical Association 52 (No.4): 466-96
}
\keyword{ package }
